requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.1
django-mongodb-engine==0.6.0
django-redis==5.4.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.6.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
drf_spectacular_websocket==1.2.9
h11==0.14.0
hyperlink==21.0.0
idna==2.7
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.1.0
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pymongo==3.12.3
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.20.0
rpds-py==0.20.0
service-identity==24.1.0
setuptools-scm==8.1.0
sqlparse==0.2.4
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.24.3
gunicorn==20.1.0
uvicorn==0.23.2
httptools==0.6.0
uvloop==0.18.0
websockets==14.1
whitenoise==6.8.2
wsproto==1.2.0
zope.interface==7.1.0
